ref(chat): Decompose agent executor into phase modules
executeAgentRunInPrivacyContext was ~2,100 lines with a dozen mutable
closure variables that existed so the catch block could persist yield,
timeout, and auth-pause continuations. Split it into chat/agent phase
modules with explicit contracts:

- request.ts owns the role-group types and routing interpretation
  (requester derivation, surface inference, destination asserts,
  session identifiers)
- resume.ts owns run resumability: safe-boundary checkpoints, durable
  input commit, resume snapshots, and translation of expected endings
  into AgentRunOutcome values, so the executor catch block is a thin
  translation plus genuine-error guards
- tools.ts wires sandbox, MCP/plugin auth orchestration, provider
  restoration, and Pi tool surfaces through a params/return contract
- prompt.ts owns user-turn content parts, steering message shape,
  resume-safe history trimming, and full prompt assembly
- skills.ts, session.ts, and sandbox.ts carry the remaining phases;
  index.ts stays the composition root and inline provider retry loop

The private flattenAgentRunRequest shim is dissolved; phases consume
the groups directly. Persisted sandbox reuse state (sandboxId,
dependency profile hash) moves from the policy group to the state
group where it belongs, with the egress trace-propagation override
staying in policy as sandboxTracePropagation.

agent-run-helpers.ts is dissolved along domain lines: Pi transcript
predicates and runtime-turn-context manipulation move to
chat/pi/transcript.ts, reply-quality text classifiers become private
to services/turn-result.ts, summarizeMessageText joins chat/logging,
and the prompt/skill helpers join their agent phase modules. Two
production-dead exports (getSuccessfulToolCalls,
prependMissingRuntimeTurnContext) are removed with their tests.

No behavior change: persistence ordering, telemetry keys, and delivery
semantics are preserved; consumer import paths move to the new module
homes.

Refs #746
