+++ packages/junior/src/chat/agent/index.ts
@@ -1,16 +1,15 @@
 /**
- * Agent turn orchestration.
+ * Agent executor.
  *
- * This module owns the Pi-facing execution boundary for one Junior turn after
- * Slack/runtime code has parsed and routed the request. It assembles prompt context,
- * restores durable Pi/session state, wires tools/MCP/auth, executes the agent,
- * and persists resumable checkpoints. Slack delivery and thread presentation
- * should stay outside this file.
+ * Composition root and execution loop for one agent run slice after
+ * runtime/ingress code has parsed and routed the request. Wires the phase
+ * modules (session restore, skills, tools, prompt, resume), runs the Pi
+ * agent with the inline provider retry loop, and translates expected run
+ * endings into `AgentRunOutcome` values. Delivery and thread presentation
+ * stay outside this module.
  */
-import { Agent, type AgentTool } from "@earendil-works/pi-agent-core";
-import type { Destination, Source } from "@sentry/junior-plugin-api";
+import { Agent } from "@earendil-works/pi-agent-core";
 import { THREAD_STATE_TTL_MS, type FileUpload } from "chat";
-import { z } from "zod";
 import { botConfig } from "@/chat/config";
 import {
   extractGenAiUsageAttributes,
@@ -21,58 +20,24 @@ import {
   serializeGenAiAttribute,
   setSpanAttributes,
   setTags,
+  summarizeMessageText,
   withSpan,
   type LogContext,
 } from "@/chat/logging";
-import { listReferenceFiles } from "@/chat/discovery";
-import {
-  buildPluginSystemPromptContributions,
-  buildSystemPrompt,
-  buildTurnContextPrompt,
-} from "@/chat/prompt";
-import {
-  extractCurrentInstructionBody,
-  renderCurrentInstruction,
-} from "@/chat/current-instruction";
-import { createUserTokenStore } from "@/chat/capabilities/factory";
-import { maybeExecuteJrRpcCustomCommand } from "@/chat/capabilities/jr-rpc-command";
 import { getConfigDefaults } from "@/chat/configuration/defaults";
-import type { ChannelConfigurationService } from "@/chat/configuration/types";
 import { SkillSandbox } from "@/chat/sandbox/skill-sandbox";
 import {
-  discoverSkills,
   findSkillByName,
   parseSkillInvocation,
   type Skill,
 } from "@/chat/skills";
-import { pluginCatalogRuntime } from "@/chat/plugins/catalog-runtime";
-import {
-  createPluginHookRunner,
-  getPluginSystemPromptContributions,
-  getPluginUserPromptContributions,
-} from "@/chat/plugins/agent-hooks";
 import { McpToolManager } from "@/chat/mcp/tool-manager";
-import {
-  inferActiveMcpProvidersFromPiMessages,
-  inferLoadedSkillNamesFromPiMessages,
-} from "@/chat/pi/derived-state";
 import type { ThreadArtifactsState } from "@/chat/state/artifacts";
-import type { ConversationPendingAuthState } from "@/chat/state/conversation";
 import {
   loadConnectedMcpProviders,
   recordToolExecutionStarted,
   recordMcpProviderConnected,
 } from "@/chat/state/session-log";
-import { createTools } from "@/chat/tools";
-import type { ToolDefinition } from "@/chat/tools/definition";
-import { toActiveMcpCatalogSummaries } from "@/chat/tools/skill/mcp-tool-summary";
-import type {
-  ImageGenerateToolDeps,
-  ToolRuntimeContext,
-  WebFetchToolDeps,
-  WebSearchToolDeps,
-} from "@/chat/tools/types";
-import { createAdvisorToolDefinitions } from "@/chat/tools/advisor/tool";
 import {
   GEN_AI_PROVIDER_NAME,
   GEN_AI_SERVER_ADDRESS,
@@ -82,39 +47,11 @@ import {
   resolveGatewayModel,
 } from "@/chat/pi/client";
 import type { PiMessage } from "@/chat/pi/messages";
+import { isAssistantMessage } from "@/chat/pi/transcript";
 import { createTracedStreamFn } from "@/chat/pi/traced-stream";
-import {
-  createSandboxExecutor,
-  type SandboxAcquiredState,
-  type SandboxExecutor,
-} from "@/chat/sandbox/sandbox";
-import type { SandboxEgressTracePropagationConfig } from "@/chat/sandbox/egress/tracing";
-import type { SandboxWorkspace } from "@/chat/sandbox/workspace";
 import { shouldEmitDevAgentTrace } from "@/chat/runtime/dev-agent-trace";
-import type { AssistantStatusSpec } from "@/chat/slack/assistant-thread/status";
-import type { SlackConversationContext } from "@/chat/slack/conversation-context";
-import {
-  createAgentTools,
-  type ToolExecutionReport,
-} from "@/chat/tools/agent-tools";
-import { mergeArtifactsState } from "@/chat/runtime/thread-state";
-import {
-  CooperativeTurnYieldError,
-  TurnInputCommitLostError,
-  isTurnInputCommitLostError,
-} from "@/chat/runtime/turn";
+import { isTurnInputCommitLostError } from "@/chat/runtime/turn";
 import type { AgentRunOutcome } from "@/chat/runtime/agent-run-outcome";
-import {
-  buildUserTurnText,
-  encodeNonImageAttachmentForPrompt,
-  getSessionIdentifiers,
-  hasRuntimeTurnContext,
-  isAssistantMessage,
-  stripRuntimeTurnContext,
-  summarizeMessageText,
-  toObservablePromptPart,
-  upsertActiveSkill,
-} from "@/chat/agent-run-helpers";
 import { buildTurnResult } from "@/chat/services/turn-result";
 import {
   isProviderRetryError,
@@ -130,25 +67,7 @@ import {
   hasAgentTurnUsage,
   type AgentTurnUsage,
 } from "@/chat/usage";
-import {
-  loadTurnSessionRecord,
-  persistAuthPauseSessionRecord,
-  persistRunningSessionRecord,
-  persistTimeoutSessionRecord,
-  persistYieldSessionRecord,
-} from "@/chat/services/turn-session-record";
-import type { AgentTurnSurface } from "@/chat/state/turn-session";
-import type { CredentialContext } from "@/chat/credentials/context";
-import { parseSlackThreadId } from "@/chat/slack/context";
-import { createMcpAuthOrchestration } from "@/chat/services/mcp-auth-orchestration";
-import { createPluginAuthOrchestration } from "@/chat/services/plugin-auth-orchestration";
-import { createPluginEgress } from "@/chat/egress/plugin";
-import { createRequester, type Requester } from "@/chat/requester";
-import {
-  AuthorizationFlowDisabledError,
-  AuthorizationPauseError,
-  type AuthorizationFlowMode,
-} from "@/chat/services/auth-pause";
+import { AuthorizationFlowDisabledError } from "@/chat/services/auth-pause";
 import {
   resolveConversationPrivacy,
   runWithConversationPrivacy,
@@ -156,6 +75,23 @@ import {
   toGenAiMessagesTraceAttributes,
   type ConversationPrivacy,
 } from "@/chat/conversation-privacy";
+import {
+  assertCorrelationDestinationMatch,
+  assertRequesterDestinationMatch,
+  getSessionIdentifiers,
+  requesterFromRouting,
+  surfaceFromRouting,
+  type AgentRunRequest,
+} from "@/chat/agent/request";
+import { restoreSessionRecord } from "@/chat/agent/session";
+import { discoverRunSkills, restoreSkillRuntime } from "@/chat/agent/skills";
+import {
+  assemblePrompt,
+  buildPromptInput,
+  buildSteeringPiMessage,
+} from "@/chat/agent/prompt";
+import { wireAgentTools } from "@/chat/agent/tools";
+import { createResumeState, type ResumeState } from "@/chat/agent/resume";
 
 const AGENT_ABORT_SETTLE_GRACE_MS = 5_000;
 
@@ -197,564 +133,68 @@ function waitForAbortSettlement(
   });
 }
 
-/** Carries the user-visible content and prior transcript for one agent-run slice. */
-export interface AgentRunInput {
-  messageText: string;
-  userAttachments?: AgentRunAttachment[];
-  inboundAttachmentCount?: number;
-  omittedImageAttachmentCount?: number;
-  /** Durable Pi transcript for this conversation, excluding ephemeral turn context. */
-  piMessages?: PiMessage[];
-  conversationContext?: string;
-}
-
-/** Carries identity and addressing needed to route tools, auth, and delivery. */
-export interface AgentRunRouting {
-  credentialContext?: CredentialContext;
-  requester?: Requester;
-  source: Source;
-  slackConversation?: SlackConversationContext;
-  destination: Destination;
-  surface?: AgentTurnSurface;
-  dispatch?: {
-    actor?: { id: string; type: string };
-    metadata?: Record<string, string>;
-    plugin?: string;
-  };
-  correlation?: {
-    conversationId?: string;
-    threadId?: string;
-    turnId?: string;
-    runId?: string;
-    channelId?: string;
-    channelName?: string;
-    teamId?: string;
-    messageTs?: string;
-    threadTs?: string;
-    requesterId?: string;
-  };
-  toolChannelId?: string;
-}
-
-/**
- * Carries execution limits, dependency overrides, and persisted sandbox
- * reuse state for one run slice.
- */
-export interface AgentRunPolicy {
-  /** Absolute wall-clock deadline for this host request, in milliseconds. */
-  turnDeadlineAtMs?: number;
-  authorizationFlowMode?: AuthorizationFlowMode;
-  configuration?: Record<string, unknown>;
-  channelConfiguration?: ChannelConfigurationService;
-  skillDirs?: string[];
-  sandbox?: {
-    sandboxId?: string;
-    sandboxDependencyProfileHash?: string;
-    /** Per-slice override for app-owned sandbox egress trace propagation. */
-    tracePropagation?: SandboxEgressTracePropagationConfig;
-  };
-  toolOverrides?: {
-    imageGenerate?: ImageGenerateToolDeps;
-    webFetch?: WebFetchToolDeps;
-    webSearch?: WebSearchToolDeps;
-  };
-}
-
-/** Carries durable state snapshots already loaded by the caller. */
-export interface AgentRunState {
-  artifactState?: ThreadArtifactsState;
-  pendingAuth?: ConversationPendingAuthState;
-}
-
-/**
- * Carries notification-only callbacks for streaming UI and status surfaces;
- * their failures never affect the run.
- */
-export interface AgentRunObservers {
-  onTextDelta?: (deltaText: string) => void | Promise<void>;
-  onAssistantMessageStart?: () => void | Promise<void>;
-  onToolInvocation?: (invocation: {
-    toolName: string;
-    params: Record<string, unknown>;
-  }) => void | Promise<void>;
-  onToolResult?: (result: ToolExecutionReport) => void | Promise<void>;
-  onStatus?: (status: AssistantStatusSpec) => void | Promise<void>;
-}
-
-/** Carries durable-worker ports that commit or update resumable run state. */
-export interface AgentRunDurability {
-  onInputCommitted?: () => void | Promise<void>;
-  /** Return true when the durable worker should pause at the next Pi boundary. */
-  shouldYield?: () => boolean;
-  drainSteeringMessages?: (
-    accept: (messages: AgentRunSteeringMessage[]) => Promise<void>,
-  ) => Promise<AgentRunSteeringMessage[]>;
-  recordPendingAuth?: (
-    pendingAuth: ConversationPendingAuthState,
-  ) => void | Promise<void>;
-  onSandboxAcquired?: (sandbox: SandboxAcquiredState) => void | Promise<void>;
-  onArtifactStateUpdated?: (
-    artifactState: ThreadArtifactsState,
-  ) => void | Promise<void>;
-}
-
-/** Groups the per-slice run request by the runtime role each field serves. */
-export interface AgentRunRequest {
-  input: AgentRunInput;
-  routing: AgentRunRouting;
-  policy?: AgentRunPolicy;
-  state?: AgentRunState;
-  observers?: AgentRunObservers;
-  durability?: AgentRunDurability;
-}
-
-type FlatAgentRunRequest = AgentRunInput &
-  AgentRunRouting &
-  AgentRunPolicy &
-  AgentRunState &
-  AgentRunObservers &
-  AgentRunDurability;
-
-/**
- * Interim shim: run internals still consume the historical flat shape
- * (grouped rewrite deferred to #746 Phase 5). The groups must stay
- * key-disjoint or later spreads silently shadow earlier fields.
- */
-function flattenAgentRunRequest(request: AgentRunRequest): FlatAgentRunRequest {
-  return {
-    ...request.input,
-    ...request.routing,
-    ...(request.policy ?? {}),
-    ...(request.state ?? {}),
-    ...(request.observers ?? {}),
-    ...(request.durability ?? {}),
-  };
-}
-
-export interface AgentRunAttachment {
-  data?: Buffer;
-  mediaType: string;
-  filename?: string;
-  promptText?: string;
-}
-
-export interface AgentRunSteeringMessage {
-  omittedImageAttachmentCount?: number;
-  text: string;
-  timestampMs?: number;
-  userAttachments?: AgentRunAttachment[];
-}
-
-let startupDiscoveryLogged = false;
-const MAX_ROUTER_ATTACHMENT_PREVIEW_CHARS = 2_000;
-
-type UserTurnContentPart =
-  | { type: "text"; text: string }
-  | { type: "image"; data: string; mimeType: string };
-
-const legacyStoredTextPartSchema = z
-  .object({
-    text: z.string(),
-    type: z.literal("text"),
-  })
-  .strict();
-
-type UserTurnAttachment = NonNullable<AgentRunInput["userAttachments"]>[number];
-
-function buildOmittedImageAttachmentNotice(count: number): string {
-  return [
-    "<omitted-image-attachments>",
-    `count: ${count}`,
-    "Slack included image attachments with this turn, but this runtime cannot analyze images because no vision model is configured.",
-    "Do not claim that no image was attached.",
-    "If the user asks about image contents, explain that image analysis is unavailable in this runtime and continue with any text or non-image files that are still available.",
-    "</omitted-image-attachments>",
-  ].join("\n");
-}
-
-function trimRouterAttachmentText(text: string): string {
-  const normalized = text.replaceAll("\0", " ").trim();
-  if (!normalized) {
-    return "";
-  }
-  return normalized.length <= MAX_ROUTER_ATTACHMENT_PREVIEW_CHARS
-    ? normalized
-    : `${normalized.slice(0, MAX_ROUTER_ATTACHMENT_PREVIEW_CHARS)}...`;
-}
-
-function extractSliceUsage(
-  messages: PiMessage[],
-  beforeMessageCount: number,
-): AgentTurnUsage | undefined {
-  const usage = extractGenAiUsageSummary(
-    ...messages.slice(beforeMessageCount).filter(isAssistantMessage),
-  );
-  return hasAgentTurnUsage(usage) ? usage : undefined;
-}
-
-function requesterFromContext(
-  context: FlatAgentRunRequest,
-): Requester | undefined {
-  return actorRequesterFromContext(context);
-}
-
-/** Reject requester identities that do not belong to the active destination. */
-function assertRequesterDestinationMatch(context: FlatAgentRunRequest): void {
-  const { destination, requester } = context;
-  if (!requester) {
-    return;
-  }
-  if (requester.platform !== destination.platform) {
-    throw new TypeError(
-      `Requester platform "${requester.platform}" does not match destination platform "${destination.platform}"`,
-    );
-  }
-  if (
-    requester.platform === "slack" &&
-    destination.platform === "slack" &&
-    requester.teamId !== destination.teamId
-  ) {
-    throw new TypeError("Slack requester team does not match destination team");
-  }
-}
-
-/** Reject legacy Slack correlation fields that conflict with the destination. */
-function assertCorrelationDestinationMatch(context: FlatAgentRunRequest): void {
-  const { correlation, destination } = context;
-  if (destination.platform !== "slack") {
-    return;
-  }
-  if (
-    correlation?.channelId !== undefined &&
-    correlation.channelId !== destination.channelId
-  ) {
-    throw new TypeError(
-      "Slack correlation channel does not match destination channel",
-    );
-  }
-  if (
-    correlation?.teamId !== undefined &&
-    correlation.teamId !== destination.teamId
-  ) {
-    throw new TypeError(
-      "Slack correlation team does not match destination team",
-    );
-  }
-}
-
-function actorRequesterFromContext(
-  context: FlatAgentRunRequest,
-): Requester | undefined {
-  return createRequester(context.requester, {
-    platform:
-      context.requester?.platform ??
-      (context.destination.platform === "slack" ? "slack" : undefined),
-    teamId:
-      (context.destination.platform === "slack"
-        ? context.destination.teamId
-        : undefined) ??
-      context.correlation?.teamId ??
-      (context.requester?.platform === "slack"
-        ? context.requester.teamId
-        : undefined),
-    userId: context.correlation?.requesterId,
-  });
-}
-
-function toolInvocationDestination(context: FlatAgentRunRequest): Destination {
-  if (context.destination.platform !== "slack" || !context.toolChannelId) {
-    return context.destination;
-  }
-  return {
-    platform: "slack",
-    teamId: context.destination.teamId,
-    channelId: context.toolChannelId,
-  };
-}
-
-function surfaceFromContext(
-  context: FlatAgentRunRequest,
-): AgentTurnSurface | undefined {
-  if (context.surface) {
-    return context.surface;
-  }
-  const conversationId =
-    context.correlation?.conversationId ??
-    context.correlation?.threadId ??
-    context.correlation?.runId;
-  if (
-    context.slackConversation ||
-    (conversationId ? parseSlackThreadId(conversationId) : undefined)
-  ) {
-    return "slack";
-  }
-  if (conversationId) {
-    return "api";
-  }
-  return undefined;
-}
-
-function supportsRouterTextPreview(mediaType: string): boolean {
-  const baseMediaType = mediaType.split(";", 1)[0]?.trim().toLowerCase();
-  if (!baseMediaType) {
-    return false;
-  }
-  return (
-    baseMediaType.startsWith("text/") ||
-    baseMediaType === "application/json" ||
-    baseMediaType === "application/xml" ||
-    baseMediaType === "application/x-www-form-urlencoded" ||
-    baseMediaType.endsWith("+json") ||
-    baseMediaType.endsWith("+xml")
-  );
-}
-
-function buildRouterAttachmentBlock(attachment: UserTurnAttachment): string {
-  if (attachment.promptText) {
-    return trimRouterAttachmentText(attachment.promptText);
-  }
-
-  const header = [
-    "<attachment>",
-    `filename: ${attachment.filename ?? "unnamed"}`,
-    `media_type: ${attachment.mediaType}`,
-  ];
-
-  if (attachment.data && supportsRouterTextPreview(attachment.mediaType)) {
-    const preview = trimRouterAttachmentText(attachment.data.toString("utf8"));
-    if (preview) {
-      return [
-        ...header,
-        "<text-preview>",
-        preview,
-        "</text-preview>",
-        "</attachment>",
-      ].join("\n");
-    }
-  }
-
-  return [...header, "</attachment>"].join("\n");
-}
-
-function buildUserTurnInput(args: {
-  omittedImageAttachmentCount: number;
-  userAttachments?: AgentRunInput["userAttachments"];
-  userTurnText: string;
-}): {
-  routerBlocks: string[];
-  userContentParts: UserTurnContentPart[];
-} {
-  const routerBlocks: string[] = [];
-  const userContentParts: UserTurnContentPart[] = [
-    { type: "text", text: args.userTurnText },
-  ];
-
-  if (args.omittedImageAttachmentCount > 0) {
-    const omittedImagesNotice = buildOmittedImageAttachmentNotice(
-      args.omittedImageAttachmentCount,
-    );
-    userContentParts.push({ type: "text", text: omittedImagesNotice });
-    routerBlocks.push(omittedImagesNotice);
-  }
-
-  for (const attachment of args.userAttachments ?? []) {
-    routerBlocks.push(buildRouterAttachmentBlock(attachment));
-
-    if (attachment.promptText) {
-      userContentParts.push({
-        type: "text",
-        text: attachment.promptText,
-      });
-      continue;
-    }
-
-    if (attachment.mediaType.startsWith("image/")) {
-      if (!attachment.data) {
-        throw new Error("Image attachment is missing image data");
-      }
-      userContentParts.push({
-        type: "image",
-        data: attachment.data.toString("base64"),
-        mimeType: attachment.mediaType,
-      });
-      continue;
-    }
-
-    if (!attachment.data) {
-      throw new Error("Attachment is missing attachment data");
-    }
-
-    userContentParts.push({
-      type: "text",
-      text: encodeNonImageAttachmentForPrompt({
-        data: attachment.data,
-        mediaType: attachment.mediaType,
-        filename: attachment.filename,
-      }),
-    });
-  }
-
-  return { routerBlocks, userContentParts };
-}
-
-/**
- * Convert a mid-run user message into the Pi user message shape used for
- * steering injection and parked-conversation session-log appends, so both
- * paths store identical durable history.
- */
-export function buildSteeringPiMessage(
-  message: AgentRunSteeringMessage,
-): PiMessage {
-  const { userContentParts } = buildUserTurnInput({
-    userTurnText: buildUserTurnText(message.text),
-    userAttachments: message.userAttachments,
-    omittedImageAttachmentCount: message.omittedImageAttachmentCount ?? 0,
-  });
-  return {
-    role: "user",
-    content: userContentParts,
-    timestamp: message.timestampMs ?? Date.now(),
-  } as PiMessage;
-}
-
-function withoutTrailingUncheckpointedUserPrompt(
-  messages: PiMessage[] | undefined,
-  userContentParts: UserTurnContentPart[],
-): PiMessage[] {
-  if (!messages || messages.length === 0) {
-    return [];
-  }
-
-  const lastMessage = messages.at(-1) as
-    | { content?: unknown; role?: unknown }
-    | undefined;
-  if (lastMessage?.role !== "user") {
-    return messages;
-  }
-  const comparableLastMessage = stripRuntimeTurnContext([
-    lastMessage as PiMessage,
-  ])[0] as { content?: unknown } | undefined;
-  if (
-    !userPromptContentMatches(comparableLastMessage?.content, userContentParts)
-  ) {
-    return messages;
-  }
-  return messages.slice(0, -1);
-}
-
-/** Match stored resume prompts against the current wrapped prompt shape. */
-function userPromptContentMatches(
-  storedContent: unknown,
-  currentContent: UserTurnContentPart[],
-): boolean {
-  if (JSON.stringify(storedContent) === JSON.stringify(currentContent)) {
-    return true;
-  }
-  if (!Array.isArray(storedContent)) {
-    return false;
-  }
-  if (storedContent.length !== currentContent.length) {
-    return false;
-  }
-
-  return storedContent.every((storedPart, index) => {
-    const currentPart = currentContent[index];
-    if (index === 0 && currentPart?.type === "text") {
-      const legacyTextPart = legacyStoredTextPartSchema.safeParse(storedPart);
-      if (legacyTextPart.success) {
-        // TODO(v0.84.0): Remove legacy unwrapped resume prompt matching after
-        // pre-current-instruction session records expire.
-        return legacyTextPartMatchesCurrentText(
-          legacyTextPart.data.text,
-          currentPart.text,
-        );
-      }
-    }
-
-    return JSON.stringify(storedPart) === JSON.stringify(currentPart);
-  });
-}
-
-function legacyTextPartMatchesCurrentText(
-  storedText: string,
-  currentText: string,
-): boolean {
-  const storedInstructionBody = extractCurrentInstructionBody(storedText);
-  if (storedInstructionBody !== undefined) {
-    return renderCurrentInstruction(storedInstructionBody) === currentText;
-  }
-
-  return renderCurrentInstruction(storedText) === currentText;
-}
-
 /** Run a full agent turn: discover skills, execute tools, and return the assistant reply. */
 export async function executeAgentRun(
   request: AgentRunRequest,
 ): Promise<AgentRunOutcome> {
-  const context = flattenAgentRunRequest(request);
-  const messageText = request.input.messageText;
   const conversationPrivacy = resolveConversationPrivacy({
-    channelId: context.correlation?.channelId,
+    channelId: request.routing.correlation?.channelId,
     conversationId:
-      context.correlation?.conversationId ??
-      context.correlation?.threadId ??
-      context.correlation?.runId,
+      request.routing.correlation?.conversationId ??
+      request.routing.correlation?.threadId ??
+      request.routing.correlation?.runId,
     // Source-confirmed visibility from the live event's channel_type; without
     // it the turn fails closed to private telemetry capture.
-    visibility: context.slackConversation?.visibility,
+    visibility: request.routing.slackConversation?.visibility,
   });
   return runWithConversationPrivacy(conversationPrivacy ?? "private", () =>
-    executeAgentRunInPrivacyContext(messageText, context, conversationPrivacy),
+    executeAgentRunInPrivacyContext(request, conversationPrivacy),
   );
 }
 
 async function executeAgentRunInPrivacyContext(
-  messageText: string,
-  context: FlatAgentRunRequest,
+  request: AgentRunRequest,
   conversationPrivacy: ConversationPrivacy | undefined,
 ): Promise<AgentRunOutcome> {
-  if (!context.destination) {
+  const { input, routing } = request;
+  const policy = request.policy ?? {};
+  const state = request.state ?? {};
+  const observers = request.observers ?? {};
+  const durability = request.durability ?? {};
+
+  if (!routing.destination) {
     throw new TypeError("Assistant reply generation requires a destination");
   }
-  assertRequesterDestinationMatch(context);
-  assertCorrelationDestinationMatch(context);
+  assertRequesterDestinationMatch(routing);
+  assertCorrelationDestinationMatch(routing);
 
   const replyStartedAtMs = Date.now();
   const configuredTurnDeadlineAtMs = replyStartedAtMs + botConfig.turnTimeoutMs;
-  const contextTurnDeadlineAtMs =
-    typeof context.turnDeadlineAtMs === "number" &&
-    Number.isFinite(context.turnDeadlineAtMs)
-      ? Math.floor(context.turnDeadlineAtMs)
+  const policyTurnDeadlineAtMs =
+    typeof policy.turnDeadlineAtMs === "number" &&
+    Number.isFinite(policy.turnDeadlineAtMs)
+      ? Math.floor(policy.turnDeadlineAtMs)
       : undefined;
   const turnDeadlineAtMs =
-    contextTurnDeadlineAtMs === undefined
+    policyTurnDeadlineAtMs === undefined
       ? configuredTurnDeadlineAtMs
-      : Math.min(configuredTurnDeadlineAtMs, contextTurnDeadlineAtMs);
+      : Math.min(configuredTurnDeadlineAtMs, policyTurnDeadlineAtMs);
   const turnTimeoutBudgetMs = Math.max(0, turnDeadlineAtMs - replyStartedAtMs);
-  let timeoutResumeConversationId: string | undefined;
-  let timeoutResumeSessionId: string | undefined;
-  let timeoutResumeSliceId = 1;
-  let timeoutResumeMessages: PiMessage[] = [];
-  let beforeMessageCount = 0;
-  let turnStartMessageIndex: number | undefined;
-  let lastKnownSandboxId: string | undefined = context.sandbox?.sandboxId;
+
+  let resume: ResumeState | undefined;
+  let lastKnownSandboxId: string | undefined = state.sandbox?.sandboxId;
   let lastKnownSandboxDependencyProfileHash: string | undefined =
-    context.sandbox?.sandboxDependencyProfileHash;
-  let loadedSkillNamesForResume: string[] = [];
+    state.sandbox?.sandboxDependencyProfileHash;
   let mcpToolManager: McpToolManager | undefined;
   let connectedMcpProviders = new Set<string>();
   let canRecordMcpProviders = false;
-  let sandboxExecutor: SandboxExecutor | undefined;
-  let timedOut = false;
-  let cooperativeYieldError: CooperativeTurnYieldError | undefined;
-  let inputCommitted = false;
   let turnUsage: AgentTurnUsage | undefined;
   let thinkingSelection: TurnThinkingSelection | undefined;
-  const requester = requesterFromContext(context);
-  const actorRequester = actorRequesterFromContext(context);
-  const surface = surfaceFromContext(context);
-  const runSource = context.source;
-  const credentialActor = context.credentialContext?.actor;
+  const requester = requesterFromRouting(routing);
+  const surface = surfaceFromRouting(routing);
+  const runSource = routing.source;
+  const userInput = input.messageText;
+  const credentialActor = routing.credentialContext?.actor;
   const credentialActorLogContext = credentialActor
     ? {
         actorType: credentialActor.type,
@@ -765,24 +205,26 @@ async function executeAgentRunInPrivacyContext(
       }
     : {};
   const sessionRecordLogContext = {
-    threadId: context.correlation?.threadId,
-    requesterId: context.correlation?.requesterId,
-    channelId: context.correlation?.channelId,
-    runId: context.correlation?.runId,
+    threadId: routing.correlation?.threadId,
+    requesterId: routing.correlation?.requesterId,
+    channelId: routing.correlation?.channelId,
+    runId: routing.correlation?.runId,
     ...credentialActorLogContext,
     assistantUserName: botConfig.userName,
     modelId: botConfig.modelId,
   };
+  const { conversationId: sessionConversationId, sessionId } =
+    getSessionIdentifiers(routing);
   const recordConnectedMcpProvider = async (provider: string) => {
     if (
       !canRecordMcpProviders ||
-      !timeoutResumeConversationId ||
+      !sessionConversationId ||
       connectedMcpProviders.has(provider)
     ) {
       return;
     }
     await recordMcpProviderConnected({
-      conversationId: timeoutResumeConversationId,
+      conversationId: sessionConversationId,
       provider,
       ttlMs: THREAD_STATE_TTL_MS,
     });
@@ -796,65 +238,32 @@ async function executeAgentRunInPrivacyContext(
       await recordConnectedMcpProvider(provider);
     }
   };
-  const getSandboxMetadata = () =>
-    sandboxExecutor
-      ? {
-          sandboxId: sandboxExecutor.getSandboxId(),
-          sandboxDependencyProfileHash:
-            sandboxExecutor.getDependencyProfileHash(),
-        }
-      : {
-          sandboxId: lastKnownSandboxId,
-          sandboxDependencyProfileHash: lastKnownSandboxDependencyProfileHash,
-        };
+  const getSandboxMetadata = () => ({
+    sandboxId: lastKnownSandboxId,
+    sandboxDependencyProfileHash: lastKnownSandboxDependencyProfileHash,
+  });
 
   try {
     const shouldTrace = shouldEmitDevAgentTrace();
     const spanContext: LogContext = {
-      conversationId:
-        context.correlation?.conversationId ??
-        context.correlation?.threadId ??
-        context.correlation?.runId,
-      slackThreadId: context.correlation?.threadId,
-      slackUserId: context.correlation?.requesterId,
-      slackChannelId: context.correlation?.channelId,
-      runId: context.correlation?.runId,
+      conversationId: sessionConversationId,
+      slackThreadId: routing.correlation?.threadId,
+      slackUserId: routing.correlation?.requesterId,
+      slackChannelId: routing.correlation?.channelId,
+      runId: routing.correlation?.runId,
       ...credentialActorLogContext,
       assistantUserName: botConfig.userName,
       modelId: botConfig.modelId,
     };
 
     // ── Skill discovery ──────────────────────────────────────────────
-    const availableSkills = await discoverSkills({
-      additionalRoots: context.skillDirs,
+    const availableSkills = await discoverRunSkills({
+      skillDirs: policy.skillDirs,
+      spanContext,
     });
-    if (!startupDiscoveryLogged) {
-      startupDiscoveryLogged = true;
-      const plugins = pluginCatalogRuntime.getProviders();
-      const roots = [
-        ...new Set(availableSkills.map((skill) => skill.skillPath)),
-      ].sort();
-      logInfo(
-        "startup_discovery_summary",
-        spanContext,
-        {
-          "app.skill.count": availableSkills.length,
-          "app.skill.names": availableSkills.map((skill) => skill.name).sort(),
-          "app.file.directories": roots,
-          "app.plugin.count": plugins.length,
-          "app.plugin.names": plugins
-            .map((plugin) => plugin.manifest.name)
-            .sort(),
-        },
-        "Discovered startup SOUL/skills/plugins",
-      );
-    }
-    let baseInstructions = "";
-    let configurationValues: Record<string, unknown>;
-    const userInput = messageText;
     if (shouldTrace) {
-      const inboundAttachmentCount = context.inboundAttachmentCount ?? 0;
-      const promptAttachmentCount = context.userAttachments?.length ?? 0;
+      const inboundAttachmentCount = input.inboundAttachmentCount ?? 0;
+      const promptAttachmentCount = input.userAttachments?.length ?? 0;
       logInfo(
         "agent_message_in",
         spanContext,
@@ -866,7 +275,7 @@ async function executeAgentRunInPrivacyContext(
           // look indistinguishable from Slack ingress dropping attachments.
           "app.message.attachment_count": inboundAttachmentCount,
           "app.message.prompt_attachment_count": promptAttachmentCount,
-          "messaging.message.id": context.correlation?.messageTs ?? "",
+          "messaging.message.id": routing.correlation?.messageTs ?? "",
         },
         "Agent message received",
       );
@@ -876,33 +285,47 @@ async function executeAgentRunInPrivacyContext(
       ? findSkillByName(skillInvocation.skillName, availableSkills)
       : null;
     const activeSkills: Skill[] = [];
+    let loadedSkillNamesForResume: string[] = [];
     const syncLoadedSkillNamesForResume = () => {
       loadedSkillNamesForResume = activeSkills.map((skill) => skill.name);
     };
     const skillSandbox = new SkillSandbox(availableSkills, activeSkills);
 
-    // ── Turn Session Record ────────────────────────────────────────
-    const { conversationId: sessionConversationId, sessionId } =
-      getSessionIdentifiers(context);
-    const turnSessionState = await loadTurnSessionRecord({
-      conversationId: sessionConversationId,
-      sessionId,
-    });
-    const { resumedFromSessionRecord, currentSliceId, existingSessionRecord } =
-      turnSessionState;
-    timeoutResumeConversationId = sessionConversationId;
-    timeoutResumeSessionId = sessionId;
-    timeoutResumeSliceId = currentSliceId;
+    // ── Turn session record ──────────────────────────────────────────
+    const {
+      sessionRecordState,
+      resumedFromSessionRecord,
+      currentSliceId,
+      existingSessionRecord,
+    } = await restoreSessionRecord(routing);
     canRecordMcpProviders = Boolean(
-      turnSessionState.canUseTurnSession && sessionConversationId && sessionId,
+      sessionRecordState.canUseTurnSession &&
+      sessionConversationId &&
+      sessionId,
     );
+    resume = createResumeState({
+      channelName: routing.correlation?.channelName,
+      destination: routing.destination,
+      durability,
+      getLoadedSkillNames: () => loadedSkillNamesForResume,
+      logContext: sessionRecordLogContext,
+      recordActiveMcpProviders,
+      requester,
+      runSource,
+      sessionConversationId,
+      sessionId,
+      sessionRecordState,
+      startedAtMs: replyStartedAtMs,
+      surface,
+    });
+    const runResume = resume;
     const recordParentToolExecutionStart = async (event: {
       args: unknown;
       toolCallId: string;
       toolName: string;
     }) => {
       if (
-        !turnSessionState.canUseTurnSession ||
+        !sessionRecordState.canUseTurnSession ||
         !sessionConversationId ||
         !sessionId
       ) {
@@ -931,176 +354,50 @@ async function executeAgentRunInPrivacyContext(
         );
       }
     };
-    const persistedConfigurationValues = context.channelConfiguration
-      ? await context.channelConfiguration.resolveValues()
+    const persistedConfigurationValues = policy.channelConfiguration
+      ? await policy.channelConfiguration.resolveValues()
       : {};
-    configurationValues = {
+    const configurationValues: Record<string, unknown> = {
       ...getConfigDefaults(),
-      ...(context.configuration ?? {}),
+      ...(policy.configuration ?? {}),
       ...persistedConfigurationValues,
     };
-    // ── Sandbox ──────────────────────────────────────────────────────
-    const authRequesterId =
-      context.credentialContext?.actor.type === "user"
-        ? context.credentialContext.actor.userId
-        : undefined;
-    const userTokenStore = createUserTokenStore();
-    const pluginHooks = createPluginHookRunner({
-      requester: actorRequester,
-    });
-    sandboxExecutor = createSandboxExecutor({
-      sandboxId: context.sandbox?.sandboxId,
-      sandboxDependencyProfileHash:
-        context.sandbox?.sandboxDependencyProfileHash,
-      traceContext: spanContext,
-      tracePropagation: context.sandbox?.tracePropagation,
-      credentialEgress: context.credentialContext,
-      agentHooks: pluginHooks,
-      onSandboxAcquired: async (sandbox) => {
-        lastKnownSandboxId = sandbox.sandboxId;
-        lastKnownSandboxDependencyProfileHash =
-          sandbox.sandboxDependencyProfileHash;
-        await context.onSandboxAcquired?.(sandbox);
-      },
-      runBashCustomCommand: async (command) => {
-        const result = await maybeExecuteJrRpcCustomCommand(command, {
-          activeSkill: skillSandbox.getActiveSkill(),
-          channelConfiguration: context.channelConfiguration,
-          requesterId: actorRequester?.userId,
-          onConfigurationValueChanged: (key, value) => {
-            if (value === undefined) {
-              delete configurationValues[key];
-              return;
-            }
-            configurationValues[key] = value;
-          },
-        });
-        return result.handled
-          ? { handled: true, result: result.result }
-          : { handled: false };
-      },
-    });
-    const currentSandboxExecutor = sandboxExecutor;
-    sandboxExecutor.configureSkills(availableSkills);
-    sandboxExecutor.configureReferenceFiles(listReferenceFiles());
+
     // Match the history source the agent will actually receive so crash retries
     // do not let an unstripped running record suppress fresh turn context.
     const priorPiMessages = resumedFromSessionRecord
       ? existingSessionRecord?.piMessages
-      : context.piMessages;
+      : input.piMessages;
     connectedMcpProviders = new Set(
-      turnSessionState.canUseTurnSession && sessionConversationId
+      sessionRecordState.canUseTurnSession && sessionConversationId
         ? await loadConnectedMcpProviders({
             conversationId: sessionConversationId,
           })
         : [],
     );
-    let sandboxPromise: Promise<SandboxWorkspace> | undefined;
-    let sandboxPromiseId: string | undefined;
-    const clearSandboxPromise = (): void => {
-      sandboxPromise = undefined;
-      sandboxPromiseId = undefined;
-    };
-    const getSandbox = (reason: {
-      trigger: string;
-      path?: string;
-      cmd?: string;
-      cwd?: string;
-    }): Promise<SandboxWorkspace> => {
-      const currentSandboxId = currentSandboxExecutor.getSandboxId();
-      if (
-        sandboxPromise &&
-        sandboxPromiseId &&
-        currentSandboxId !== sandboxPromiseId
-      ) {
-        clearSandboxPromise();
-      }
-
-      if (!sandboxPromise) {
-        logInfo(
-          "sandbox_boot_requested",
-          spanContext,
-          {
-            "app.sandbox.boot.trigger": reason.trigger,
-            ...(reason.path ? { "file.path": reason.path } : {}),
-            ...(reason.cmd ? { "process.executable.name": reason.cmd } : {}),
-            ...(reason.cwd ? { "file.directory": reason.cwd } : {}),
-          },
-          "Lazy sandbox boot requested",
-        );
-        sandboxPromise = currentSandboxExecutor
-          .createSandbox()
-          .then((sandbox) => {
-            sandboxPromiseId = sandbox.sandboxId;
-            return sandbox;
-          })
-          .catch((error) => {
-            clearSandboxPromise();
-            throw error;
-          });
-      }
-      return sandboxPromise;
-    };
-    const sandbox: SandboxWorkspace = {
-      readFileToBuffer: async (input) =>
-        (
-          await getSandbox({
-            trigger: "workspace.readFileToBuffer",
-            path: input.path,
-          })
-        ).readFileToBuffer(input),
-      runCommand: async (input) =>
-        (
-          await getSandbox({
-            trigger: "workspace.runCommand",
-            cmd: input.cmd,
-            cwd: input.cwd,
-          })
-        ).runCommand(input),
-    };
 
     // ── Restore skill runtime handles from durable Pi history ────────
-    for (const skillName of inferLoadedSkillNamesFromPiMessages(
+    await restoreSkillRuntime({
+      activeSkills,
+      invokedSkill,
       priorPiMessages,
-    )) {
-      const restoredSkill = await skillSandbox.loadSkill(skillName);
-      if (restoredSkill) {
-        upsertActiveSkill(activeSkills, restoredSkill);
-        syncLoadedSkillNamesForResume();
-      }
-    }
-    if (invokedSkill) {
-      const restoredSkill = await skillSandbox.loadSkill(invokedSkill.name);
-      if (restoredSkill) {
-        upsertActiveSkill(activeSkills, restoredSkill);
-        syncLoadedSkillNamesForResume();
-      }
-    }
-
-    const promptConversationContext =
-      context.piMessages && context.piMessages.length > 0
-        ? undefined
-        : context.conversationContext;
-    const userTurnText = buildUserTurnText(
-      userInput,
-      promptConversationContext,
-    );
-    const { routerBlocks, userContentParts } = buildUserTurnInput({
-      omittedImageAttachmentCount: context.omittedImageAttachmentCount ?? 0,
-      userAttachments: context.userAttachments,
-      userTurnText,
+      skillSandbox,
+      syncLoadedSkillNamesForResume,
     });
+
+    // ── Prompt input ─────────────────────────────────────────────────
+    const { routerBlocks, userContentParts } = buildPromptInput(input);
     const preAgentPromptMessages = (): PiMessage[] =>
-      existingSessionRecord?.piMessages ?? [...(context.piMessages ?? [])];
+      existingSessionRecord?.piMessages ?? [...(input.piMessages ?? [])];
 
     thinkingSelection = await selectTurnThinkingLevel({
       completeObject,
-      conversationContext: context.conversationContext,
+      conversationContext: input.conversationContext,
       context: {
-        threadId: context.correlation?.threadId,
-        channelId: context.correlation?.channelId,
-        requesterId: context.correlation?.requesterId,
-        runId: context.correlation?.runId,
+        threadId: routing.correlation?.threadId,
+        channelId: routing.correlation?.channelId,
+        requesterId: routing.correlation?.requesterId,
+        runId: routing.correlation?.runId,
       },
       currentTurnBlocks: routerBlocks,
       fastModelId: botConfig.fastModelId,
@@ -1118,462 +415,100 @@ async function executeAgentRunInPrivacyContext(
     });
 
     // ── Mutable turn state ───────────────────────────────────────────
-    timeoutResumeMessages = [];
     const generatedFiles: FileUpload[] = [];
     const replyFiles: FileUpload[] = [];
     const artifactStatePatch: Partial<ThreadArtifactsState> = {};
     const toolCalls: string[] = [];
-    let advisorTools: AgentTool[] = [];
     let agent: Agent | undefined;
-    let latestSafeBoundaryMessages: PiMessage[] = [];
-    const getResumeSnapshot = (): PiMessage[] => {
-      const currentMessages = agent ? [...agent.state.messages] : [];
-      return latestSafeBoundaryMessages.length > currentMessages.length
-        ? [...latestSafeBoundaryMessages]
-        : currentMessages;
-    };
+    const currentAgentMessages = (): PiMessage[] =>
+      agent ? [...agent.state.messages] : [];
 
-    // ── MCP auth orchestration ───────────────────────────────────────
-    const slackDestination =
-      context.destination.platform === "slack"
-        ? context.destination
-        : undefined;
-    const slackChannelId = slackDestination?.channelId;
-
-    const mcpAuth = createMcpAuthOrchestration({
-      abortAgent: () => agent?.abort(),
-      conversationId: sessionConversationId,
-      sessionId,
-      requesterId: authRequesterId,
-      channelId: slackChannelId,
-      destination: context.destination,
-      source: runSource,
-      threadTs: context.correlation?.threadTs,
-      toolChannelId: context.toolChannelId,
-      userMessage: userInput,
-      pendingAuth: context.pendingAuth,
-      getConfiguration: () => configurationValues,
-      getArtifactState: () => context.artifactState,
-      getMergedArtifactState: () =>
-        mergeArtifactsState(context.artifactState ?? {}, artifactStatePatch),
-      recordPendingAuth: context.recordPendingAuth,
-      authorizationFlowMode: context.authorizationFlowMode,
-    });
-    const pluginAuth = createPluginAuthOrchestration({
-      abortAgent: () => agent?.abort(),
-      conversationId: sessionConversationId,
-      sessionId,
-      requesterId: authRequesterId,
-      channelId: slackChannelId,
-      destination: context.destination,
-      source: runSource,
-      threadTs: context.correlation?.threadTs,
-      userMessage: userInput,
-      channelConfiguration: context.channelConfiguration,
-      pendingAuth: context.pendingAuth,
-      recordPendingAuth: context.recordPendingAuth,
-      authorizationFlowMode: context.authorizationFlowMode,
-      userTokenStore,
-    });
-
-    mcpToolManager = new McpToolManager(
-      pluginCatalogRuntime.getMcpProviders(),
-      {
-        authProviderFactory: mcpAuth.authProviderFactory,
-        onAuthorizationRequired: mcpAuth.onAuthorizationRequired,
-      },
-    );
-    const turnMcpToolManager = mcpToolManager;
-    const getPendingAuthPause = () =>
-      pluginAuth.getPendingPause() ?? mcpAuth.getPendingPause();
     setTags({
       conversationId: spanContext.conversationId,
-      slackThreadId: context.correlation?.threadId,
-      slackUserId: context.correlation?.requesterId,
-      slackChannelId: context.correlation?.channelId,
-      runId: context.correlation?.runId,
+      slackThreadId: routing.correlation?.threadId,
+      slackUserId: routing.correlation?.requesterId,
+      slackChannelId: routing.correlation?.channelId,
+      runId: routing.correlation?.runId,
       ...credentialActorLogContext,
       assistantUserName: botConfig.userName,
       modelId: botConfig.modelId,
     });
 
-    // ── Tool creation ────────────────────────────────────────────────
-    const loadableSkills = availableSkills.filter(
-      (skill) =>
-        skill.disableModelInvocation !== true ||
-        skill.name === invokedSkill?.name,
-    );
-    const commonToolRuntimeContext = {
-      conversationId: sessionConversationId,
-      userText: userInput,
-      artifactState: context.artifactState,
-      configuration: configurationValues,
-      egress: createPluginEgress({
-        credentialContext: context.credentialContext,
-        pluginAuth: {
-          async handleAuthRequired(signal) {
-            await pluginAuth.maybeHandleAuthSignal({
-              auth_required: {
-                ...(signal.authorization
-                  ? { authorization: signal.authorization }
-                  : {}),
-                createdAtMs: Date.now(),
-                grant: signal.grant,
-                kind: signal.kind,
-                message: signal.message,
-                provider: signal.provider,
-              },
-            });
-          },
-        },
-      }),
-      mcpToolManager: turnMcpToolManager,
-      sandbox,
+    // ── Tool wiring ──────────────────────────────────────────────────
+    const wiring = await wireAgentTools({
+      abortAgent: () => agent?.abort(),
+      activeSkills,
+      actorRequester: requester,
+      artifactStatePatch,
+      availableSkills,
+      configurationValues,
+      connectedMcpProviders,
+      conversationPrivacy,
+      durability,
+      generatedFiles,
+      invokedSkill,
+      observers,
+      onSandboxMetadataChanged: (sandbox) => {
+        lastKnownSandboxId = sandbox.sandboxId;
+        lastKnownSandboxDependencyProfileHash =
+          sandbox.sandboxDependencyProfileHash;
+      },
+      policy,
+      preAgentPromptMessages,
+      priorPiMessages,
+      recordConnectedMcpProvider,
+      replyFiles,
+      resume: runResume,
+      routing,
+      sessionConversationId,
+      sessionId,
+      skillSandbox,
+      spanContext,
+      state,
       surface,
-      advisor: {
-        config: botConfig.advisor,
-        conversationId: sessionConversationId,
-        conversationPrivacy,
-        parentSessionId: sessionId,
-        logContext: spanContext,
-        getTools: () => advisorTools,
-        streamFn: createTracedStreamFn({ conversationPrivacy }),
-      },
-    };
-    const toolSource = runSource;
-    const toolDestination = toolInvocationDestination(context);
-    let toolRuntimeContext: ToolRuntimeContext;
-    if (toolSource.platform === "slack") {
-      if (toolDestination.platform !== "slack") {
-        throw new TypeError("Slack tool runtime requires a Slack destination");
-      }
-      toolRuntimeContext = {
-        ...commonToolRuntimeContext,
-        destination: toolDestination,
-        requester:
-          actorRequester?.platform === "slack" ? actorRequester : undefined,
-        source: toolSource,
-      };
-    } else {
-      if (toolDestination.platform !== "local") {
-        throw new TypeError("Local tool runtime requires a local destination");
-      }
-      toolRuntimeContext = {
-        ...commonToolRuntimeContext,
-        destination: toolDestination,
-        requester:
-          actorRequester?.platform === "local" ? actorRequester : undefined,
-        source: toolSource,
-      };
-    }
-    const tools = createTools(
-      loadableSkills,
-      {
-        getGeneratedFile: (filename) =>
-          generatedFiles.find((file) => file.filename === filename),
-        onGeneratedArtifactFiles: (files) => {
-          generatedFiles.push(...files);
-        },
-        onGeneratedFiles: (files) => {
-          replyFiles.push(...files);
-        },
-        onArtifactStatePatch: async (patch) => {
-          Object.assign(artifactStatePatch, patch);
-          await context.onArtifactStateUpdated?.(
-            mergeArtifactsState(
-              context.artifactState ?? {},
-              artifactStatePatch,
-            ),
-          );
-        },
-        toolOverrides: context.toolOverrides,
-        onSkillLoaded: async (loadedSkill) => {
-          const resolvedSkill = await skillSandbox.loadSkill(loadedSkill.name);
-          const effective = resolvedSkill ?? loadedSkill;
-          upsertActiveSkill(activeSkills, effective);
-          syncLoadedSkillNamesForResume();
-          if (await turnMcpToolManager.activateForSkill(effective)) {
-            await recordConnectedMcpProvider(effective.pluginProvider!);
-          }
-          if (mcpAuth.getPendingPause()) {
-            // Auth pause requested — suppress loadSkill failure and let the
-            // aborted turn park cleanly.
-            return undefined;
-          }
-          if (!effective.pluginProvider) {
-            return undefined;
-          }
-          if (
-            !turnMcpToolManager
-              .getActiveProviders()
-              .includes(effective.pluginProvider)
-          ) {
-            return undefined;
-          }
-          const availableToolCount = turnMcpToolManager.getActiveToolCatalog({
-            provider: effective.pluginProvider,
-          }).length;
-          return {
-            mcp_provider: effective.pluginProvider,
-            available_tool_count: availableToolCount,
-          };
-        },
-      },
-      toolRuntimeContext,
-    );
-
-    const toolGuidance = Object.entries(
-      tools as Record<string, ToolDefinition<any>>,
-    ).map(([name, definition]) => ({
-      name,
-      promptGuidelines: definition.promptGuidelines,
-      promptSnippet: definition.promptSnippet,
-    }));
-
-    // ── MCP provider activation ──────────────────────────────────────
-    // If a prior turn left an MCP provider pending user authorization, skip
-    // eager restoration of that provider here. Without this guard, a later
-    // unrelated turn in the same conversation can try to activate the
-    // still-unauthenticated provider, throw McpAuthorizationPauseError, and
-    // abort before the agent sees the user's request.
-    //
-    // Skipping only suppresses the eager-restore path. The agent can still
-    // trigger the auth flow intentionally (via loadSkill + searchMcpTools)
-    // when the user's request genuinely requires that provider.
-    const pendingMcpProvider =
-      context.pendingAuth?.kind === "mcp"
-        ? context.pendingAuth.provider
-        : undefined;
-
-    // Restore providers visible in durable Pi session history. In serverless
-    // runtimes, later slices and follow-up turns usually run in a fresh
-    // process, so in-memory MCP clients cannot be reused.
-    const providersToRestore = new Set([
-      ...connectedMcpProviders,
-      ...inferActiveMcpProvidersFromPiMessages(priorPiMessages),
-    ]);
-    for (const provider of providersToRestore) {
-      if (provider === pendingMcpProvider) {
-        continue; // awaiting user authorization — skip to avoid aborting unrelated turns
-      }
-      if (await turnMcpToolManager.activateProvider(provider)) {
-        await recordConnectedMcpProvider(provider);
-      }
-      if (mcpAuth.getPendingPause()) {
-        timeoutResumeMessages = preAgentPromptMessages();
-        throw mcpAuth.getPendingPause()!;
-      }
-    }
-    // Activate MCP for skills recovered from durable Pi history.
-    for (const skill of activeSkills) {
-      if (skill.pluginProvider === pendingMcpProvider) {
-        continue; // awaiting user authorization — skip to avoid aborting unrelated turns
-      }
-      if (await turnMcpToolManager.activateForSkill(skill)) {
-        await recordConnectedMcpProvider(skill.pluginProvider!);
-      }
-      if (mcpAuth.getPendingPause()) {
-        timeoutResumeMessages = preAgentPromptMessages();
-        throw mcpAuth.getPendingPause()!;
-      }
-    }
+      syncLoadedSkillNamesForResume,
+      toolCalls,
+      userInput,
+    });
+    mcpToolManager = wiring.mcpToolManager;
+    const sandboxExecutor = wiring.sandboxExecutor;
+    const getPendingAuthPause = wiring.getPendingAuthPause;
 
     // ── Prompt context ───────────────────────────────────────────────
-    const activeMcpCatalogs = toActiveMcpCatalogSummaries(
-      turnMcpToolManager.getActiveToolCatalog(),
-    );
-    const hasPromptCheckpoint =
-      resumedFromSessionRecord &&
-      existingSessionRecord?.turnStartMessageIndex !== undefined;
-    const shouldPromptAgent = !resumedFromSessionRecord || !hasPromptCheckpoint;
-    // Every re-prompt shape must trim a trailing checkpointed copy of the same
-    // user prompt, including redelivery of the same inbound message after a
-    // lost input commit against a still-`running` record; otherwise the prompt
-    // appears twice in Pi history.
-    const promptHistoryMessages = shouldPromptAgent
-      ? withoutTrailingUncheckpointedUserPrompt(
-          priorPiMessages,
-          userContentParts,
-        )
-      : existingSessionRecord!.piMessages;
-    const needsBootstrapContextForPrompt =
-      shouldPromptAgent && !hasRuntimeTurnContext(promptHistoryMessages);
-    const systemPromptContributions =
-      await getPluginSystemPromptContributions(toolSource);
-    const pluginSystemPrompt = buildPluginSystemPromptContributions(
-      systemPromptContributions,
-    );
-    baseInstructions = [
-      buildSystemPrompt({ source: toolSource }),
-      pluginSystemPrompt,
-    ]
-      .filter((section): section is string => Boolean(section))
-      .join("\n\n");
-    const pluginUserPromptContributions = !shouldPromptAgent
-      ? []
-      : await getPluginUserPromptContributions({
-          context: toolRuntimeContext,
-        });
-    const turnContextPrompt =
-      needsBootstrapContextForPrompt || pluginUserPromptContributions.length > 0
-        ? buildTurnContextPrompt({
-            availableSkills,
-            activeMcpCatalogs,
-            includeSessionContext: needsBootstrapContextForPrompt,
-            pluginPromptContributions: pluginUserPromptContributions,
-            toolGuidance,
-            runtime: {
-              conversationId: spanContext.conversationId,
-              slackConversation: context.slackConversation,
-            },
-            dispatch: context.dispatch
-              ? {
-                  ...context.dispatch,
-                  destination: context.destination,
-                  source: toolSource,
-                }
-              : undefined,
-            invocation: skillInvocation,
-            requester: actorRequester,
-            artifactState: context.artifactState,
-            configuration: configurationValues,
-          })
-        : null;
-    const turnContextParts: UserTurnContentPart[] = turnContextPrompt
-      ? [{ type: "text", text: turnContextPrompt }]
-      : [];
-    const promptContentParts: UserTurnContentPart[] = [
-      ...turnContextParts,
-      ...userContentParts,
-    ];
-
-    const inputMessages = [
-      {
-        role: "system",
-        content: [{ type: "text", text: baseInstructions }],
-      },
-      {
-        role: "user",
-        content: promptContentParts.map((part) => toObservablePromptPart(part)),
-      },
-    ];
-    const inputMessagesAttribute = serializeGenAiAttribute(
-      conversationPrivacy !== "public"
-        ? inputMessages.map(toGenAiMessageMetadata)
-        : inputMessages,
-    );
-
-    // ── Agent tools ──────────────────────────────────────────────────
-    const onToolCall = async (
-      toolName: string,
-      params: Record<string, unknown>,
-    ) => {
-      toolCalls.push(toolName);
-      try {
-        await context.onToolInvocation?.({ toolName, params });
-      } catch (error) {
-        logWarn(
-          "tool_invocation_observer_failed",
-          spanContext,
-          {
-            "gen_ai.tool.name": toolName,
-            "exception.message":
-              error instanceof Error ? error.message : String(error),
-          },
-          "Tool invocation observer failed",
-        );
-      }
-    };
-    const agentTools = createAgentTools(
-      tools,
-      skillSandbox,
-      spanContext,
-      context.onStatus,
-      sandboxExecutor,
-      pluginAuth,
-      onToolCall,
-      pluginHooks,
+    const {
+      baseInstructions,
+      inputMessages,
+      inputMessagesAttribute,
+      promptContentParts,
+      promptHistoryMessages,
+      shouldPromptAgent,
+    } = await assemblePrompt({
+      activeMcpCatalogs: wiring.activeMcpCatalogs,
+      actorRequester: requester,
+      artifactState: state.artifactState,
+      availableSkills,
+      configurationValues,
       conversationPrivacy,
-      context.onToolResult,
-    );
-    advisorTools = createAgentTools(
-      createAdvisorToolDefinitions(tools),
-      skillSandbox,
+      existingSessionPiMessages: existingSessionRecord?.piMessages,
+      existingTurnStartMessageIndex:
+        existingSessionRecord?.turnStartMessageIndex,
+      invocation: skillInvocation,
+      priorPiMessages,
+      resumedFromSessionRecord,
+      routing,
       spanContext,
-      context.onStatus,
-      sandboxExecutor,
-      pluginAuth,
-      onToolCall,
-      pluginHooks,
-      conversationPrivacy,
-      context.onToolResult,
-    );
-    // Keep Pi's native tool schema static for the whole turn. Ideally this
-    // would use provider-native tool loading/search APIs, but Pi's generic
-    // AgentTool surface cannot yet express OpenAI/Anthropic deferred MCP tools.
-    // Until it can, MCP tools are searched/disclosed as data and executed
-    // through callMcpTool so provider cache/session affinity never sees a
-    // mid-run native tool-list mutation.
+      toolGuidance: wiring.toolGuidance,
+      toolRuntimeContext: wiring.toolRuntimeContext,
+      userContentParts,
+    });
 
     // ── Agent execution ──────────────────────────────────────────────
     let hasEmittedText = false;
     let needsSeparator = false;
-    const commitInput = async (): Promise<void> => {
-      if (inputCommitted) {
-        return;
-      }
-      await context.onInputCommitted?.();
-      inputCommitted = true;
-    };
-    const persistSafeBoundary = async (
-      messages: PiMessage[],
-    ): Promise<boolean> => {
-      if (
-        !turnSessionState.canUseTurnSession ||
-        !sessionConversationId ||
-        !sessionId
-      ) {
-        return false;
-      }
-
-      const persisted = await persistRunningSessionRecord({
-        channelName: context.correlation?.channelName,
-        conversationId: sessionConversationId,
-        destination: context.destination,
-        source: runSource,
-        sessionId,
-        sliceId: currentSliceId,
-        messages,
-        loadedSkillNames: loadedSkillNamesForResume,
-        logContext: sessionRecordLogContext,
-        requester,
-        ...(surface ? { surface } : {}),
-        ...(turnStartMessageIndex !== undefined
-          ? { turnStartMessageIndex }
-          : {}),
-      });
-      if (!persisted) {
-        return false;
-      }
-
-      latestSafeBoundaryMessages = [...messages];
-      return true;
-    };
-    const requireDurableInputCheckpoint = async (
-      messages: PiMessage[],
-    ): Promise<boolean> => {
-      const persisted = await persistSafeBoundary(messages);
-      if (!persisted && context.onInputCommitted) {
-        throw new TurnInputCommitLostError(
-          `Durable turn input could not be checkpointed for conversation=${sessionConversationId ?? "unknown"} session=${sessionId ?? "unknown"}`,
-        );
-      }
-      return persisted;
-    };
     const drainSteeringMessages = async (): Promise<void> => {
       if (
-        !context.drainSteeringMessages ||
-        !turnSessionState.canUseTurnSession ||
+        !durability.drainSteeringMessages ||
+        !sessionRecordState.canUseTurnSession ||
         !sessionConversationId ||
         !sessionId
       ) {
@@ -1582,12 +517,12 @@ async function executeAgentRunInPrivacyContext(
 
       try {
         let steeredMessageCount = 0;
-        await context.drainSteeringMessages(async (messages) => {
+        await durability.drainSteeringMessages(async (messages) => {
           const piMessages = messages.map(buildSteeringPiMessage);
           if (piMessages.length === 0) {
             return;
           }
-          await requireDurableInputCheckpoint([
+          await runResume.requireDurableInputCheckpoint([
             ...agent!.state.messages,
             ...piMessages,
           ]);
@@ -1621,19 +556,6 @@ async function executeAgentRunInPrivacyContext(
         );
       }
     };
-    const yieldAtSafeBoundaryIfDue = (): void => {
-      if (!context.shouldYield?.()) {
-        return;
-      }
-
-      timeoutResumeMessages = getResumeSnapshot();
-      cooperativeYieldError = new CooperativeTurnYieldError(
-        `Agent turn yielded at a safe boundary after ${
-          Date.now() - replyStartedAtMs
-        }ms`,
-      );
-      throw cooperativeYieldError;
-    };
 
     const apiKeyOverride = getPiGatewayApiKey();
     agent = new Agent({
@@ -1642,14 +564,14 @@ async function executeAgentRunInPrivacyContext(
       steeringMode: "all",
       prepareNextTurn: async () => {
         await drainSteeringMessages();
-        yieldAtSafeBoundaryIfDue();
+        runResume.yieldAtSafeBoundaryIfDue(currentAgentMessages());
         return undefined;
       },
       initialState: {
         systemPrompt: baseInstructions,
         model: resolveGatewayModel(botConfig.modelId),
         thinkingLevel: toAgentThinkingLevel(thinkingSelection.thinkingLevel),
-        tools: agentTools,
+        tools: wiring.agentTools,
       },
     });
 
@@ -1658,22 +580,24 @@ async function executeAgentRunInPrivacyContext(
         return recordParentToolExecutionStart(event);
       }
       if (event.type === "turn_end" && event.toolResults.length > 0) {
-        return persistSafeBoundary([...agent!.state.messages]).then(
-          () => undefined,
-        );
+        return runResume
+          .persistSafeBoundary([...agent!.state.messages])
+          .then(() => undefined);
       }
       if (event.type === "message_start") {
-        Promise.resolve(context.onAssistantMessageStart?.()).catch((error) => {
-          logWarn(
-            "streaming_message_start_error",
-            {},
-            {
-              "exception.message":
-                error instanceof Error ? error.message : String(error),
-            },
-            "Failed to deliver assistant message start to stream coordinator",
-          );
-        });
+        Promise.resolve(observers.onAssistantMessageStart?.()).catch(
+          (error) => {
+            logWarn(
+              "streaming_message_start_error",
+              {},
+              {
+                "exception.message":
+                  error instanceof Error ? error.message : String(error),
+              },
+              "Failed to deliver assistant message start to stream coordinator",
+            );
+          },
+        );
         if (hasEmittedText) {
           needsSeparator = true;
         }
@@ -1688,7 +612,7 @@ async function executeAgentRunInPrivacyContext(
       needsSeparator = false;
       hasEmittedText = true;
 
-      Promise.resolve(context.onTextDelta?.(text)).catch((error) => {
+      Promise.resolve(observers.onTextDelta?.(text)).catch((error) => {
         logWarn(
           "streaming_text_delta_error",
           {},
@@ -1702,19 +626,20 @@ async function executeAgentRunInPrivacyContext(
     });
 
     let newMessages: PiMessage[] = [];
-    beforeMessageCount = agent.state.messages.length;
     try {
       if (resumedFromSessionRecord) {
         agent.state.messages = shouldPromptAgent
           ? promptHistoryMessages
           : existingSessionRecord!.piMessages;
-        turnStartMessageIndex = existingSessionRecord!.turnStartMessageIndex;
+        runResume.setTurnStartMessageIndex(
+          existingSessionRecord!.turnStartMessageIndex,
+        );
       } else if (promptHistoryMessages.length > 0) {
         agent.state.messages = [...promptHistoryMessages];
       }
-      beforeMessageCount = agent.state.messages.length;
+      runResume.setBeforeMessageCount(agent.state.messages.length);
       if (shouldPromptAgent) {
-        turnStartMessageIndex = beforeMessageCount;
+        runResume.setTurnStartMessageIndex(agent.state.messages.length);
       }
 
       await withSpan(
@@ -1729,12 +654,13 @@ async function executeAgentRunInPrivacyContext(
             timestamp: Date.now(),
           } as PiMessage;
           if (shouldPromptAgent) {
-            const promptPersisted = await requireDurableInputCheckpoint([
-              ...agent.state.messages,
-              freshPromptMessage,
-            ]);
+            const promptPersisted =
+              await runResume.requireDurableInputCheckpoint([
+                ...agent!.state.messages,
+                freshPromptMessage,
+              ]);
             if (promptPersisted) {
-              await commitInput();
+              await runResume.commitInput();
             }
           }
 
@@ -1744,8 +670,8 @@ async function executeAgentRunInPrivacyContext(
             let timeoutId: ReturnType<typeof setTimeout> | undefined;
             const timeoutPromise = new Promise<never>((_, reject) => {
               const rejectWithTimeout = () => {
-                timedOut = true;
-                agent.abort();
+                runResume.markTimedOut();
+                agent!.abort();
                 reject(
                   new Error(
                     `Agent turn timed out after ${turnTimeoutBudgetMs}ms`,
@@ -1763,7 +689,7 @@ async function executeAgentRunInPrivacyContext(
             try {
               return await Promise.race([run, timeoutPromise]);
             } catch (error) {
-              if (timedOut) {
+              if (runResume.timedOut) {
                 logWarn(
                   "agent_turn_timeout",
                   {},
@@ -1800,10 +726,14 @@ async function executeAgentRunInPrivacyContext(
                     "Timed-out agent run did not settle after abort before resume snapshot",
                   );
                 }
-                timeoutResumeMessages = getResumeSnapshot();
+                runResume.captureResumeSnapshot(
+                  runResume.getResumeSnapshot(currentAgentMessages()),
+                );
               }
               if (getPendingAuthPause()) {
-                timeoutResumeMessages = getResumeSnapshot();
+                runResume.captureResumeSnapshot(
+                  runResume.getResumeSnapshot(currentAgentMessages()),
+                );
                 throw getPendingAuthPause()!;
               }
               throw error;
@@ -1815,16 +745,18 @@ async function executeAgentRunInPrivacyContext(
           };
 
           let run = shouldPromptAgent
-            ? agent.prompt(freshPromptMessage)
-            : agent.continue();
+            ? agent!.prompt(freshPromptMessage)
+            : agent!.continue();
           let retryUsage: AgentTurnUsage | undefined;
           for (let attempt = 0; ; attempt += 1) {
             promptResult = await runAgentStep(run);
-            if (cooperativeYieldError) {
-              throw cooperativeYieldError;
+            if (runResume.cooperativeYieldError) {
+              throw runResume.cooperativeYieldError;
             }
 
-            newMessages = agent.state.messages.slice(beforeMessageCount);
+            newMessages = agent!.state.messages.slice(
+              runResume.beforeMessageCount,
+            );
             const outputMessages = newMessages.filter(isAssistantMessage);
             const outputMessagesAttribute = serializeGenAiAttribute(
               conversationPrivacy !== "public"
@@ -1833,7 +765,7 @@ async function executeAgentRunInPrivacyContext(
             );
             const usageSummary = extractGenAiUsageSummary(
               promptResult,
-              agent.state,
+              agent!.state,
               ...outputMessages,
             );
             const currentUsage = hasAgentTurnUsage(usageSummary)
@@ -1851,7 +783,9 @@ async function executeAgentRunInPrivacyContext(
               ...extractGenAiUsageAttributes(usageSummary),
             });
             if (getPendingAuthPause()) {
-              timeoutResumeMessages = getResumeSnapshot();
+              runResume.captureResumeSnapshot(
+                runResume.getResumeSnapshot(currentAgentMessages()),
+              );
               throw getPendingAuthPause()!;
             }
 
@@ -1859,15 +793,15 @@ async function executeAgentRunInPrivacyContext(
             const providerRetry = nextProviderRetry({
               attempt,
               lastAssistant,
-              messages: agent.state.messages,
+              messages: agent!.state.messages,
             });
             if (!providerRetry) {
               break;
             }
 
             retryUsage = turnUsage;
-            agent.state.messages = providerRetry.messages;
-            await persistSafeBoundary(providerRetry.messages);
+            agent!.state.messages = providerRetry.messages;
+            await runResume.persistSafeBoundary(providerRetry.messages);
             logWarn(
               "agent_turn_provider_retry",
               spanContext,
@@ -1875,7 +809,7 @@ async function executeAgentRunInPrivacyContext(
               "Retrying transient provider failure",
             );
             await sleep(providerRetry.delayMs);
-            run = agent.continue();
+            run = agent!.continue();
           }
         },
         {
@@ -1892,9 +826,7 @@ async function executeAgentRunInPrivacyContext(
             ? { "app.ai.session.conversation_id": sessionConversationId }
             : {}),
           ...(sessionId ? { "app.ai.turn.session_id": sessionId } : {}),
-          ...(timeoutResumeSliceId
-            ? { "app.ai.turn.slice_id": timeoutResumeSliceId }
-            : {}),
+          ...(currentSliceId ? { "app.ai.turn.slice_id": currentSliceId } : {}),
           "app.ai.reasoning_effort": thinkingSelection.thinkingLevel,
           ...toGenAiMessagesTraceAttributes("app.ai.input", inputMessages),
           ...(inputMessagesAttribute
@@ -1907,7 +839,7 @@ async function executeAgentRunInPrivacyContext(
     }
 
     if (
-      turnSessionState.canUseTurnSession &&
+      sessionRecordState.canUseTurnSession &&
       sessionConversationId &&
       sessionId
     ) {
@@ -1929,9 +861,9 @@ async function executeAgentRunInPrivacyContext(
         replyFiles,
         artifactStatePatch,
         toolCalls,
-        sandboxId: currentSandboxExecutor.getSandboxId(),
+        sandboxId: sandboxExecutor.getSandboxId(),
         sandboxDependencyProfileHash:
-          currentSandboxExecutor.getDependencyProfileHash(),
+          sandboxExecutor.getDependencyProfileHash(),
         piMessages: [...agent.state.messages],
         durationMs: Date.now() - replyStartedAtMs,
         generatedFileCount: generatedFiles.length,
@@ -1939,114 +871,18 @@ async function executeAgentRunInPrivacyContext(
         spanContext,
         usage: turnUsage,
         thinkingSelection,
-        correlation: context.correlation,
+        correlation: routing.correlation,
         assistantUserName: botConfig.userName,
       }),
     };
   } catch (error) {
-    if (
-      cooperativeYieldError &&
-      error instanceof CooperativeTurnYieldError &&
-      timeoutResumeConversationId &&
-      timeoutResumeSessionId
-    ) {
-      turnUsage =
-        turnUsage ??
-        extractSliceUsage(timeoutResumeMessages, beforeMessageCount);
-      await recordActiveMcpProviders();
-      const sessionRecord = await persistYieldSessionRecord({
-        channelName: context.correlation?.channelName,
-        conversationId: timeoutResumeConversationId,
-        destination: context.destination,
-        source: runSource,
-        sessionId: timeoutResumeSessionId,
-        currentSliceId: timeoutResumeSliceId,
-        currentDurationMs: Date.now() - replyStartedAtMs,
+    if (resume) {
+      const { outcome } = await resume.translateExpectedEnding({
         currentUsage: turnUsage,
-        messages: timeoutResumeMessages,
-        errorMessage: error.message,
-        loadedSkillNames: loadedSkillNamesForResume,
-        logContext: sessionRecordLogContext,
-        requester,
-        ...(surface ? { surface } : {}),
+        error,
       });
-      if (!sessionRecord) {
-        throw new Error(
-          `Failed to persist cooperative yield continuation for conversation=${timeoutResumeConversationId} session=${timeoutResumeSessionId}`,
-        );
-      }
-      return { status: "suspended", resumeVersion: sessionRecord.version };
-    }
-
-    if (timedOut && timeoutResumeConversationId && timeoutResumeSessionId) {
-      turnUsage =
-        turnUsage ??
-        extractSliceUsage(timeoutResumeMessages, beforeMessageCount);
-      await recordActiveMcpProviders();
-      const sessionRecord = await persistTimeoutSessionRecord({
-        channelName: context.correlation?.channelName,
-        conversationId: timeoutResumeConversationId,
-        destination: context.destination,
-        source: runSource,
-        sessionId: timeoutResumeSessionId,
-        currentSliceId: timeoutResumeSliceId,
-        currentDurationMs: Date.now() - replyStartedAtMs,
-        currentUsage: turnUsage,
-        messages: timeoutResumeMessages,
-        errorMessage: error instanceof Error ? error.message : String(error),
-        loadedSkillNames: loadedSkillNamesForResume,
-        logContext: sessionRecordLogContext,
-        requester,
-        ...(surface ? { surface } : {}),
-      });
-      if (!sessionRecord) {
-        throw new Error(
-          `Failed to persist timeout continuation for conversation=${timeoutResumeConversationId} session=${timeoutResumeSessionId}`,
-        );
-      }
-      if (sessionRecord.state === "awaiting_resume") {
-        return { status: "suspended", resumeVersion: sessionRecord.version };
-      }
-      throw new Error(
-        sessionRecord.errorMessage ??
-          (error instanceof Error ? error.message : String(error)),
-      );
-    }
-
-    // ── MCP auth pause → session continuation ────────────────────────
-    if (
-      error instanceof AuthorizationPauseError &&
-      timeoutResumeConversationId &&
-      timeoutResumeSessionId
-    ) {
-      if (!turnUsage && timeoutResumeMessages.length > 0) {
-        turnUsage = extractSliceUsage(
-          timeoutResumeMessages,
-          beforeMessageCount,
-        );
-      }
-      await recordActiveMcpProviders();
-      const sessionRecord = await persistAuthPauseSessionRecord({
-        channelName: context.correlation?.channelName,
-        conversationId: timeoutResumeConversationId,
-        destination: context.destination,
-        source: runSource,
-        sessionId: timeoutResumeSessionId,
-        currentSliceId: timeoutResumeSliceId,
-        currentDurationMs: Date.now() - replyStartedAtMs,
-        currentUsage: turnUsage,
-        messages: timeoutResumeMessages,
-        errorMessage: error.message,
-        loadedSkillNames: loadedSkillNamesForResume,
-        logContext: sessionRecordLogContext,
-        requester,
-        ...(surface ? { surface } : {}),
-      });
-      if (sessionRecord) {
-        return {
-          status: "awaiting_auth",
-          providerDisplayName: error.providerDisplayName,
-        };
+      if (outcome) {
+        return outcome;
       }
     }
 
@@ -2059,7 +895,7 @@ async function executeAgentRunInPrivacyContext(
     if (error instanceof AuthorizationFlowDisabledError) {
       throw error;
     }
-    if (context.onInputCommitted && !inputCommitted) {
+    if (durability.onInputCommitted && !resume?.inputCommitted) {
       throw error;
     }
 
@@ -2067,10 +903,10 @@ async function executeAgentRunInPrivacyContext(
       error,
       "assistant_reply_generation_failed",
       {
-        slackThreadId: context.correlation?.threadId,
-        slackUserId: context.correlation?.requesterId,
-        slackChannelId: context.correlation?.channelId,
-        runId: context.correlation?.runId,
+        slackThreadId: routing.correlation?.threadId,
+        slackUserId: routing.correlation?.requesterId,
+        slackChannelId: routing.correlation?.channelId,
+        runId: routing.correlation?.runId,
         ...credentialActorLogContext,
         assistantUserName: botConfig.userName,
         modelId: botConfig.modelId,
