+++ packages/junior/src/chat/services/context-compaction.ts
@@ -25,7 +25,7 @@ import { logWarn, setSpanAttributes } from "@/chat/logging";
 import {
   stripRuntimeTurnContext,
   trimTrailingAssistantMessages,
-} from "@/chat/agent-run-helpers";
+} from "@/chat/pi/transcript";
 import { updateConversationStats } from "@/chat/services/conversation-memory";
 
 const RETAINED_USER_MESSAGE_TOKENS = 20_000;
