+++ packages/junior/src/chat/services/provider-retry.ts
@@ -3,7 +3,7 @@ import type { PiMessage } from "@/chat/pi/messages";
 import {
   getPiMessageRole,
   trimTrailingAssistantMessages,
-} from "@/chat/agent-run-helpers";
+} from "@/chat/pi/transcript";
 
 const PROVIDER_RETRY_DELAYS_MS = [2_000, 4_000, 8_000] as const;
 const PROVIDER_ERROR_PREFIX = "AI provider error:";
