+++ packages/junior/src/chat/services/turn-result.ts
@@ -1,6 +1,6 @@
 import type { FileUpload } from "chat";
 import { botConfig } from "@/chat/config";
-import { logInfo, logWarn } from "@/chat/logging";
+import { logInfo, logWarn, summarizeMessageText } from "@/chat/logging";
 import type { LogContext } from "@/chat/logging";
 import {
   containsNoReplyMarker,
@@ -21,13 +21,87 @@ import {
   extractAssistantText,
   getTerminalAssistantMessages,
   isAssistantMessage,
-  isExecutionEscapeResponse,
-  isRawToolPayloadResponse,
   isToolResultError,
   isToolResultMessage,
   normalizeToolNameFromResult,
-  summarizeMessageText,
-} from "@/chat/agent-run-helpers";
+} from "@/chat/pi/transcript";
+
+function isExecutionDeferralResponse(text: string): boolean {
+  return /\b(want me to proceed|do you want me to proceed|shall i proceed|can i proceed|should i proceed|let me do that now|give me a moment|tag me again|fresh invocation)\b/i.test(
+    text,
+  );
+}
+
+function isToolAccessDisclaimerResponse(text: string): boolean {
+  return /\b(i (don't|do not) have access to (active )?tool|tool results came back empty|prior results .* empty|cannot access .*tool|need to (run|load) .*tool .* first)\b/i.test(
+    text,
+  );
+}
+
+/** True when the model produced an escape response instead of executing. */
+function isExecutionEscapeResponse(text: string): boolean {
+  const trimmed = text.trim();
+  if (!trimmed) return false;
+  return (
+    isExecutionDeferralResponse(trimmed) ||
+    isToolAccessDisclaimerResponse(trimmed)
+  );
+}
+
+function parseJsonCandidate(text: string): unknown {
+  const trimmed = text.trim();
+  if (!trimmed) return undefined;
+
+  try {
+    return JSON.parse(trimmed) as unknown;
+  } catch {
+    const fenced = trimmed.match(/^```(?:json)?\s*([\s\S]*?)\s*```$/i);
+    if (!fenced) return undefined;
+    try {
+      return JSON.parse(fenced[1]) as unknown;
+    } catch {
+      return undefined;
+    }
+  }
+}
+
+function isToolPayloadShape(payload: unknown): boolean {
+  if (!payload || typeof payload !== "object") return false;
+  const record = payload as Record<string, unknown>;
+
+  const type = typeof record.type === "string" ? record.type.toLowerCase() : "";
+  if (type.startsWith("tool-")) return true;
+  if (
+    type === "tool_use" ||
+    type === "tool_call" ||
+    type === "tool_result" ||
+    type === "tool_error"
+  )
+    return true;
+
+  const hasToolName =
+    typeof record.toolName === "string" || typeof record.name === "string";
+  const hasToolInput =
+    Object.prototype.hasOwnProperty.call(record, "input") ||
+    Object.prototype.hasOwnProperty.call(record, "args");
+  if (hasToolName && hasToolInput) return true;
+
+  return false;
+}
+
+/** Detect responses that are raw tool payloads leaked as text. */
+function isRawToolPayloadResponse(text: string): boolean {
+  const parsed = parseJsonCandidate(text);
+  if (Array.isArray(parsed)) {
+    return parsed.some((entry) => isToolPayloadShape(entry));
+  }
+  if (isToolPayloadShape(parsed)) {
+    return true;
+  }
+
+  const compact = text.replace(/\s+/g, " ");
+  return /"type"\s*:\s*"tool[-_](use|call|result|error)"/i.test(compact);
+}
 
 const POST_CANVAS_REPLY_MAX_CHARS = 700;
 const POST_CANVAS_REPLY_MAX_LINES = 8;
