+++ packages/junior/src/handlers/mcp-oauth-callback.ts
@@ -376,11 +376,11 @@ async function resumeAuthorizedMcpTurn(args: {
           policy: {
             configuration: authSession.configuration,
             channelConfiguration: lockedChannelConfiguration,
-            sandbox: getPersistedSandboxState(lockedState),
           },
           state: {
             artifactState: lockedArtifacts,
             pendingAuth: lockedPendingAuth,
+            sandbox: getPersistedSandboxState(lockedState),
           },
           durability: {
             recordPendingAuth: async (nextPendingAuth) => {
